switched to component model
