+++ website/app/components/doc/component-api/index.gts
@@ -1,13 +1,25 @@
 /**
- * Copyright IBM Corp. 2021, 2025
+ * Copyright IBM Corp. 2021, 2026
  * SPDX-License-Identifier: MPL-2.0
  */
-import type { TemplateOnlyComponent } from '@ember/component/template-only';
-import { hash } from '@ember/helper';
 
+import Component from '@glimmer/component';
+import { hash } from '@ember/helper';
 import DocComponentApiProperty from 'website/components/doc/component-api/property';
+import { getDocComponentApiProperties } from 'website/shared/component-api-manifest';
+
+import type {
+  DocComponentApiProperty as DocComponentApiPropertyShape,
+  DocComponentApiSection,
+} from 'website/shared/component-api-manifest';
 
 interface DocComponentApiSignature {
+  Args: {
+    component?: string;
+    section?: DocComponentApiSection;
+    name?: string;
+    properties?: DocComponentApiPropertyShape[];
+  };
   Blocks: {
     default: [
       {
@@ -18,11 +30,72 @@ interface DocComponentApiSignature {
   Element: HTMLDivElement;
 }
 
-const DocComponentApi: TemplateOnlyComponent<DocComponentApiSignature> =
+export default class DocComponentApi extends Component<DocComponentApiSignature> {
+  get hasComponent(): boolean {
+    return (
+      typeof this.args.component === 'string' && this.args.component.length > 0
+    );
+  }
+
+  get hasRuntimeApiData(): boolean {
+    return this.hasComponent || Array.isArray(this.args.properties);
+  }
+
+  get apiProperties(): DocComponentApiPropertyShape[] {
+    if (Array.isArray(this.args.properties) === true) {
+      return this.args.properties;
+    }
+
+    const component = this.args.component;
+
+    if (typeof component !== 'string' || component.length === 0) {
+      return [];
+    }
+
+    return getDocComponentApiProperties(
+      component,
+      this.args.section ?? 'api',
+      this.args.name,
+    );
+  }
+
   <template>
     <div class="doc-component-api" ...attributes>
-      {{yield (hash Property=DocComponentApiProperty)}}
+      {{#if this.hasRuntimeApiData}}
+        {{#each this.apiProperties as |property|}}
+          <DocComponentApiProperty
+            @name={{property.name}}
+            @type={{property.type}}
+            @required={{property.required}}
+            @values={{property.values}}
+            @default={{property.default}}
+            @valueNote={{property.valueNote}}
+            @description={{property.description}}
+            @remarks={{property.remarks}}
+            @links={{property.links}}
+          >
+            {{#if property.properties}}
+              <div class="doc-component-api">
+                {{#each property.properties as |nestedProperty|}}
+                  <DocComponentApiProperty
+                    @name={{nestedProperty.name}}
+                    @type={{nestedProperty.type}}
+                    @required={{nestedProperty.required}}
+                    @values={{nestedProperty.values}}
+                    @default={{nestedProperty.default}}
+                    @valueNote={{nestedProperty.valueNote}}
+                    @description={{nestedProperty.description}}
+                    @remarks={{nestedProperty.remarks}}
+                    @links={{nestedProperty.links}}
+                  />
+                {{/each}}
+              </div>
+            {{/if}}
+          </DocComponentApiProperty>
+        {{/each}}
+      {{else}}
+        {{yield (hash Property=DocComponentApiProperty)}}
+      {{/if}}
     </div>
-  </template>;
-
-export default DocComponentApi;
+  </template>
+}
