+++ website/docs/components/alert/partials/code/component-api.md
@@ -2,7 +2,7 @@
 
 ### Alert
 
-<!-- hds-api::api -->
+<Doc::ComponentApi @component="alert" @section="api" />
 
 ### Contextual components
