+++ website/lib/markdown/markdown-process-includes.js
@@ -17,11 +17,6 @@ const fs = require('fs-extra');
 const path = require('path');
 const walkSync = require('walk-sync');
 
-const {
-  hydrateManifestApiDirectives,
-} = require('./component-api-from-manifest.js');
-const { getManifestPath } = require('./component-api-manifest-loader.js');
-
 // https://regex101.com/r/tesSl8/1
 const includeRegex = new RegExp(/^\s*@include "(.*\.md)"\s*$/, 'gm');
 
@@ -59,8 +54,7 @@ class MarkdownProcessIncludes extends Multifilter {
       // PROCESS MARKDOWN
       // we replace the `@include` directive with the actual content of the included file
       //
-      const manifestFiles = getManifestPath();
-      const includedFiles = [manifestFiles];
+      const includedFiles = [];
       let newMarkdownFileContent;
       if (markdownFileContent.match(includeRegex)) {
         newMarkdownFileContent = markdownFileContent.replace(
@@ -73,11 +67,6 @@ class MarkdownProcessIncludes extends Multifilter {
       } else {
         newMarkdownFileContent = markdownFileContent;
       }
-
-      newMarkdownFileContent = hydrateManifestApiDirectives(
-        newMarkdownFileContent,
-        inputFile,
-      );
 
       if (fs.existsSync(outputFolder)) {
         if (fs.existsSync(fullOutputPath)) {
